Janky bluetooth reconnection logic, still does not wait between socket sends, but the manual stepper works if you AutoHome and update from device() in between.  Still needs main UI and step/store programming.

+++ CamSliderCommander/SliderAxisPositionInfo.cs
@@ -8,6 +8,6 @@ namespace CamSliderCommander
 {
     public class SliderAxisPositionInfo
     {
-        public int Position { get; set; }
+        public decimal Position { get; set; }
     }
 }
